Factored entity helper methods into separate, non-db-specific helper class
