+++ src/test/java/ca/uptoeleven/status/db/DBTestHelpers.java
@@ -1,9 +1,7 @@
 package ca.uptoeleven.status.db;
 
-import ca.uptoeleven.status.core.*;
 import org.skife.jdbi.v2.DBI;
 
-import java.util.ArrayList;
 import java.util.UUID;
 
 public class DBTestHelpers {
@@ -22,26 +20,6 @@ public class DBTestHelpers {
 
 	public static String newId() {
 		return UUID.randomUUID().toString();
-	}
-
-	public static Service newServiceForTest() {
-		return Service.createNew("name", "description");
-	}
-
-	public static Incident newIncidentWithUpdateForTest() {
-		return Incident.newIncident("title", "updateDescription", IncidentState.Unplanned.INVESTIGATING, ServiceStatus.OK.getId(), new ArrayList<>());
-	}
-
-	public static Incident newIncidentForTest() {
-		return Incident.newIncident("title", "notUsed", IncidentState.Unplanned.INVESTIGATING, ServiceStatus.OK.getId(), new ArrayList<>()).withIncidentUpdatesList(new ArrayList<>());
-	}
-
-	public static Incident newPlannedIncidentForTest() {
-		return Incident.newPlannedIncident("title", "notUsed", IncidentState.Planned.STARTED, ServiceStatus.OK.getId(), new ArrayList<>()).withIncidentUpdatesList(new ArrayList<>());
-	}
-
-	public static IncidentUpdate newIncidentUpdateForTest() {
-		return IncidentUpdate.createNew("foo", IncidentState.Unplanned.INVESTIGATING, ServiceStatus.OK.getId());
 	}
 
 }
