+++ src/test/java/ca/uptoeleven/status/db/IncidentsRepositoryTest.java
@@ -6,6 +6,8 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.skife.jdbi.v2.DBI;
 
+import static ca.uptoeleven.status.core.EntityHelpers.newIncidentForTest;
+import static ca.uptoeleven.status.core.EntityHelpers.newIncidentWithUpdateForTest;
 import static ca.uptoeleven.status.db.DBTestHelpers.*;
 import static org.junit.Assert.assertEquals;
